requirements.txt
streamlit
reazonspeech
moviepy
boto3
python-dotenv
